Update synapse connected tests
